+++ packages/js/data/src/payment-gateways/types.ts
@@ -12,6 +12,7 @@ export type SettingDefinition = {
 	tip: string;
 	type: string;
 	value: string;
+	is_dismissed: string;
 };
 
 export type PaymentGateway = {
@@ -23,7 +24,8 @@ export type PaymentGateway = {
 	method_title: string;
 	method_description: string;
 	method_supports: string[];
-	settings: Record< string, SettingDefinition >;
+	settings: SettingDefinition;
+	settings_url: string;
 };
 
 export type PluginsState = {