+++ plugins/woocommerce/src/Admin/Notes/DeprecatedNotes.php
@@ -518,27 +518,6 @@ class WC_Admin_Notes_Selling_Online_Courses extends DeprecatedClassFacade {
 }
 
 /**
- * WC_Admin_Notes_Set_Up_Additional_Payment_Types.
- *
- * @deprecated since 1.7.0, use SetUpAdditionalPaymentTypes
- */
-class WC_Admin_Notes_Set_Up_Additional_Payment_Types extends DeprecatedClassFacade {
-	/**
-	 * The name of the non-deprecated class that this facade covers.
-	 *
-	 * @var string
-	 */
-	protected static $facade_over_classname = 'Automattic\WooCommerce\Internal\Admin\Notes\SetUpAdditionalPaymentTypes';
-
-	/**
-	 * The version that this class was deprecated in.
-	 *
-	 * @var string
-	 */
-	protected static $deprecated_in_version = '1.7.0';
-}
-
-/**
  * WC_Admin_Notes_Test_Checkout.
  *
  * @deprecated since 1.7.0, use TestCheckout